LayoutTests/accessibility/ios-simulator/accessibility-make-first-responder-expected.txt

Test that assistiveTechnologyMakeFirstResponder works correctly on iOS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS isWebContentFirstResponder is true
Resigning web content first respondership
PASS isWebContentFirstResponder is false
Performing accessibility object focus
PASS isWebContentFirstResponder is true
PASS successfullyParsed is true

TEST COMPLETE
